backend/requirements.txt
# Core
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-multipart==0.0.17
python-dotenv==1.0.1
pydantic==2.10.3
pydantic-settings==2.6.1
aiofiles==24.1.0

# Computer Vision & Image Processing
mediapipe==0.10.14  # Using compatible version for Python 3.11
opencv-python==4.10.0.84
Pillow==11.0.0
numpy==1.26.4

# Image Optimization
tinify==1.7.0
# Note: oxipng and pngquant are system dependencies installed via Docker/apt (fallback)

# Testing
pytest==8.3.4
pytest-asyncio==0.25.0
pytest-cov==6.0.0
httpx==0.28.1

# Code Quality
black==24.10.0
flake8==7.1.1
mypy==1.14.0
isort==5.13.2

# Monitoring & Logging
python-json-logger==3.2.0
prometheus-client==0.21.1

# Development
ipython==7.34.0  # Compatible with Python 3.9
watchdog==6.0.0